pyproject.toml
[project]
name = "skills-tooling"
version = "0.1.0"
description = "Tooling for validating and releasing skills in this repository."
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "ty",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
